Fix tests, add an example of an accession that should be ignored by ingest: INSDC001.1
