+++ ingest/tests/test_data_cchf/exclude_biosample_accessions.tsv
@@ -1,0 +1,1 @@
+SAMEA131392145
