+++ ingest/tests/test_data_cchf/exclude_insdc_accessions.tsv
@@ -1,0 +1,1 @@
+INSDC001.1
