+++ ingest/tests/test_data_cchf/metadata_post_extract.tsv
@@ -5,3 +5,4 @@ KX013464.1			PARTIAL	1	Russia: Astrakhan	Europe								Ixodoidea		297308	true			
 KX096703.1			PARTIAL	1	Kazakhstan: Sairam district	Asia								Hyalomma anatolicum		176092	true				2015	tick pool #134			513				1		2016-04-30T00:00:00Z	GenBank		Public Health England, Research		Deryabin,P.,Atshabar,B.,Sansyzbaev,Y.,Berezin,V.,Nurmakhanov,T.,Yeskhojayev,O.,Vilkova,A.,Shevtsov,A.,Hewson,R.,Atkinson,B.	2016-04-30T00:00:00Z								Orthonairovirus haemorrhagiae		3052518
 KX013483.1			PARTIAL	1	Uganda	Africa								Homo sapiens		9606	true				1958	Nakiwogo	blood		12098				1		2016-12-07T00:00:00Z	GenBank		Chumakov Institute of Poliomyelitis and Viral Encephalitides	Russia	Lukashev,A.N.,Klimentov,A.S.,Smirnova,S.E.,Dzagurova,T.K.,Drexler,J.F.,Gmyl,A.P.	2016-12-07T00:00:00Z								Orthonairovirus haemorrhagiae		3052518
 KX013485.1			PARTIAL	1	Uganda	Africa								Homo sapiens		9606	true				1958	Nakiwogo	blood		1571				1		2016-12-07T00:00:00Z	GenBank		Chumakov Institute of Poliomyelitis and Viral Encephalitides	Russia	Lukashev,A.N.,Klimentov,A.S.,Smirnova,S.E.,Dzagurova,T.K.,Drexler,J.F.,Gmyl,A.P.	2016-12-07T00:00:00Z								Orthonairovirus haemorrhagiae		3052518
+INSDC001.1			PARTIAL	1	Uganda	Africa								Homo sapiens		9606	true				1958	Nakiwogo	blood		1571				1		2016-12-07T00:00:00Z	GenBank		Chumakov Institute of Poliomyelitis and Viral Encephalitides	Russia	Lukashev,A.N.,Klimentov,A.S.,Smirnova,S.E.,Dzagurova,T.K.,Drexler,J.F.,Gmyl,A.P.	2016-12-07T00:00:00Z								Orthonairovirus haemorrhagiae		3052518
